Move no_exit test into main initialize test and rename
