+++ dask_mpi/tests/test_initialize.py
@@ -30,3 +30,14 @@ def test_small_world(mpirun):
 
     p.communicate()
     assert p.returncode != 0
+
+
+def test_no_exit(mpirun):
+    script_file = os.path.join(
+        os.path.dirname(os.path.realpath(__file__)), "initialize_no_exit.py"
+    )
+
+    p = subprocess.Popen(mpirun + ["-np", "4", sys.executable, script_file])
+
+    p.communicate()
+    assert p.returncode == 0
